\name{m3eReplication-package}
\alias{m3eReplication-package}
\alias{m3eReplication}
\title{Replication of the Static and Dynamic M3E Model}
\description{
Reproduces the matrices, equilibria, simulations and charts of the supplied
static and dynamic three-equation macroeconomic model workbooks.
}
\docType{package}

